fix: analysis endpoint fully working — tested with real AAPL data
- earnings_quality_score: int → float (Pydantic rejected 72.5)
- Added TypeError fallback for compute_yieldiq_score
- Added traceback logging in analysis router
- Removed debug prints

Verified: AAPL returns fair_value=$187.57, price=$260.48, MoS=-28%,
verdict=overvalued, Piotroski=9/9, Moat=Wide(83), all 15 engines fired.

+++ backend/models/responses.py
@@ -51,7 +51,7 @@ class QualityOutput(BaseModel):
     piotroski_score: int = 0
     piotroski_grade: str = ""
     earnings_quality_grade: str = ""
-    earnings_quality_score: int = 0
+    earnings_quality_score: float = 0
     moat: Literal["Wide", "Narrow", "None"] = "None"
     moat_score: int = 0
     momentum_score: int = 0
